Andrew: parallel fitting of subjects
Attached updated files now allow parallel
fitting of subjects. Needed a function added to
fitter.R and of course a file to run the fits
(NLMpLBAmA.R). Also realized the code at the
start of  test-models.R that reads in data
and makes sleep.RData didnt convert RT to
seconds so that is now done and attached.