Andrew: initial fits of model A
I tried to run fits of mA to the data, which mostly worked OK
except P012 and P017 caused nlm fits to crash (parameter
values heading off to infinity), but simplex fits of these
ran OK (the usual finding that simplex is more robust).
In playing with this I tweaked mA to have lower = 1e-6 for
Vs and V, which effects only PSO and DE, but did similar
with the other fitters by changing these two parameters to
a log scale. Not now sure that was necessary (some bad
fits had big negative V and Vs but after this fix I still
got some crashes although less) but left it in.

I stored the output in the attached pLBAmAnlm.RData (note
the fits are actually a mix of simplex for the first 8 subjects
and nlm for the last 6, as I ran them in chunks trying to track
down which subjects caused the crash when I ran in parallel).

Next I tried PSO. That seemed to work fine but rather slow as
you would expect. I am having trouble with DEoptim because it
crashes R on my mac. It works fine on linux.